skin_classification/requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
Django==2.2.2
Flask==1.0.3
Flask-Cors==3.0.8
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.8.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4
wrapt==1.11.2
